fix: createdAt/Cid-type 수정

+++ src/main/java/com/github/backendpart/repository/OptionRepository.java
@@ -3,5 +3,5 @@ package com.github.backendpart.repository;
 import com.github.backendpart.web.entity.OptionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OptionRepository extends JpaRepository<OptionEntity, Integer> {
+public interface OptionRepository extends JpaRepository<OptionEntity, Long> {
 }
